+++ hopp/simulation/layout/flicker_mismatch_grid.py
@@ -1,5 +1,5 @@
-from hopp.layout.flicker_mismatch import *
-from hopp.layout.shadow_flicker import create_turbines_in_grid, get_turbine_grid_shadow
+from hopp.simulation.layout.flicker_mismatch import *
+from hopp.simulation.layout.shadow_flicker import create_turbines_in_grid, get_turbine_grid_shadow
 
 from hopp.utils.log import flicker_logger as logger
 sys.path.append('.')
