+++ tests/hopp/test_hybrid.py
@@ -3,7 +3,7 @@ from pytest import approx, fixture, raises
 from pathlib import Path
 
 from hopp.simulation.sites import SiteInfo, flatirons_site
-from hopp.layout.hybrid_layout import WindBoundaryGridParameters, PVGridParameters
+from hopp.simulation.layout.hybrid_layout import WindBoundaryGridParameters, PVGridParameters
 from hopp.simulation import HybridSimulation
 from hopp.simulation.technologies import DetailedPVPlant
 from examples.Detailed_PV_Layout.detailed_pv_layout import DetailedPVParameters, DetailedPVLayout
@@ -11,7 +11,7 @@ from examples.Detailed_PV_Layout.detailed_pv_config import PVLayoutConfig
 import PySAM.Singleowner as Singleowner
 from hopp.simulation.technologies import Grid
 from hopp.utils.keys import set_nrel_key_dot_env
-from hopp.layout.pv_design_utils import size_electrical_parameters
+from hopp.simulation.layout.pv_design_utils import size_electrical_parameters
 from copy import deepcopy
 import numpy as np
 import json
